+++ noj-core/scripts/noj.ts
@@ -19,7 +19,7 @@ import { HelpCommand } from "@cliffy/command/help";
 import { CompletionsCommand } from "@cliffy/command/completions";
 import { runMigrations } from "../src/db/migrate.ts";
 import { ensureRootUser } from "../src/domains/identity/index.ts";
-import { ensureRbacSeeds } from "../src/services/seed/seed-rbac.ts";
+import { ensureRbacSeeds } from "../src/domains/system/index.ts";
 import {
   ensureAdminFromEnv,
   ensureBootstrapAdmin,
@@ -27,7 +27,7 @@ import {
   seedJudgeImages,
   seedLlmQuotas,
   seedTags,
-} from "../src/services/seed/seed-system.ts";
+} from "../src/domains/system/index.ts";
 import { importProblemBundle } from "../src/domains/catalog/index.ts";
 import { isValidTemplateFileName } from "../src/types/problem-bundle.ts";
 import { ROOT_USER_ID } from "../src/lib/constants.ts";
