+++ noj-core/src/domains/submission/services/submissions/submissions-crud.ts
@@ -54,7 +54,7 @@ import {
   isRetryableJudgeQueueError,
   pushJudgeTask,
 } from "../../../../mq/producer.ts";
-import { validateJudgeImageWithKind } from "../../../../services/judge-images.ts";
+import { validateJudgeImageWithKind } from "../../../system/index.ts";
 import { assertContestSubmissionLimit } from "../../../contest/index.ts";
 import { getStorageProvider } from "../../../../lib/storage/mod.ts";
 import { getPendingQueueSnapshot, getSubmissionQueueStatus } from "../queue.ts";
