+++ noj-core/src/domains/system/index.ts
@@ -1,2 +1,6 @@
-export * from "../../services/audit-log.ts";
-export * from "../../services/system-settings.ts";
+export * from "./services/audit-log.ts";
+export * from "./services/system-settings.ts";
+export * from "./services/announcements.ts";
+export * from "./services/judge-images.ts";
+export * from "./services/seed/seed-rbac.ts";
+export * from "./services/seed/seed-system.ts";
