+++ noj-core/src/domains/system/routes/admin-announcements.ts
@@ -19,12 +19,12 @@
  */
 
 import { Hono } from "hono";
-import { authMiddleware } from "../middleware/auth.ts";
-import { parseJsonBody } from "../lib/request.ts";
-import { getClientIp } from "../lib/rate-limit-env.ts";
-import { assertPermission } from "../lib/permissions.ts";
-import { runWithContext } from "../lib/requestContext.ts";
-import { parsePagination } from "../lib/pagination.ts";
+import { authMiddleware } from "../../../middleware/auth.ts";
+import { parseJsonBody } from "../../../lib/request.ts";
+import { getClientIp } from "../../../lib/rate-limit-env.ts";
+import { assertPermission } from "../../../lib/permissions.ts";
+import { runWithContext } from "../../../lib/requestContext.ts";
+import { parsePagination } from "../../../lib/pagination.ts";
 import {
   createAnnouncement,
   deleteAnnouncement,
