+++ noj-core/src/domains/system/routes/admin-settings.ts
@@ -1,12 +1,12 @@
 import { Hono } from "hono";
-import type { AuthEnv } from "../../middleware/auth.ts";
-import { parseJsonBody } from "../../lib/request.ts";
-import { BadRequestError } from "../../lib/errors.ts";
+import type { AuthEnv } from "../../../middleware/auth.ts";
+import { parseJsonBody } from "../../../lib/request.ts";
+import { BadRequestError } from "../../../lib/errors.ts";
 import {
   listSettings,
   resetSetting,
   updateSetting,
-} from "../../services/system-settings.ts";
+} from "../services/system-settings.ts";
 
 /**
  * 管理端系统设置路由（issue #99，挂载前缀 /api/v1/admin，见 admin/index.ts）。
