+++ noj-core/src/domains/system/routes/announcements.ts
@@ -12,10 +12,10 @@
  */
 
 import { Hono } from "hono";
-import { authMiddleware } from "../middleware/auth.ts";
-import { Channels, onEvent } from "../lib/event-bus.ts";
-import { createSseStream } from "../lib/sse-stream.ts";
-import { parsePagination } from "../lib/pagination.ts";
+import { authMiddleware } from "../../../middleware/auth.ts";
+import { Channels, onEvent } from "../../../lib/event-bus.ts";
+import { createSseStream } from "../../../lib/sse-stream.ts";
+import { parsePagination } from "../../../lib/pagination.ts";
 import {
   getPublicAnnouncement,
   listPublicAnnouncements,
