+++ noj-core/src/lib/jwt.ts
@@ -1,5 +1,5 @@
 import { type JWTPayload, jwtVerify, SignJWT } from "jose";
-import { getSetting } from "../services/system-settings.ts";
+import { getSetting } from "../domains/system/index.ts";
 
 /**
  * JWT 签发者与接收者标识。
