+++ noj-core/tests/00_migrate_test.ts
@@ -8,7 +8,7 @@
  */
 import { runMigrations } from "../src/db/migrate.ts";
 import { ensureRootUser } from "../src/domains/identity/index.ts";
-import { ensureRbacSeeds } from "../src/services/seed/seed-rbac.ts";
+import { ensureRbacSeeds } from "../src/domains/system/index.ts";
 import { getDb } from "../src/db/connection.ts";
 import { setupSchemaForTest } from "./_setup.ts";
 import { judgeImages } from "../src/db/schema.ts";
