+++ noj-core/tests/lib/xff-trust.test.ts
@@ -13,11 +13,11 @@ import { Hono } from "hono";
 import { resetDbForTest } from "../../src/db/connection.ts";
 import { getClientIp } from "../../src/lib/rate-limit-env.ts";
 
-import { updateSetting } from "../../src/services/system-settings.ts";
+import { updateSetting } from "../../src/domains/system/index.ts";
 import {
   _resetSystemSettingsForTest,
   initSystemSettings,
-} from "../../src/services/system-settings.ts";
+} from "../../src/domains/system/index.ts";
 import {
   _resetEnvSnapshotForTest,
   snapshotEnv,
