+++ noj-core/tests/routes/admin-trainings.test.ts
@@ -12,7 +12,7 @@ import {
 } from "../../src/db/schema.ts";
 import { signToken } from "../../src/lib/jwt.ts";
 import { initRedisForTest, jsonRequest } from "../lib/helper.ts";
-import { ensureRbacSeeds } from "../../src/services/seed/seed-rbac.ts";
+import { ensureRbacSeeds } from "../../src/domains/system/index.ts";
 
 await resetDbForTest();
 await initRedisForTest();
