+++ noj-core/tests/routes/messages.test.ts
@@ -20,7 +20,7 @@ import {
   _resetSystemSettingsForTest,
   initSystemSettings,
   updateSetting,
-} from "../../src/services/system-settings.ts";
+} from "../../src/domains/system/index.ts";
 import {
   enterTestContext,
   leaveTestContext,
