+++ noj-core/tests/routes/support-package.test.ts
@@ -8,7 +8,7 @@ import { createApp } from "../../src/app.ts";
 import { signToken } from "../../src/lib/jwt.ts";
 import { getDb, resetDbForTest } from "../../src/db/connection.ts";
 import { problems, roles, userRoles, users } from "../../src/db/schema.ts";
-import { ensureRbacSeeds } from "../../src/services/seed/seed-rbac.ts";
+import { ensureRbacSeeds } from "../../src/domains/system/index.ts";
 import { eq, sql } from "drizzle-orm";
 import { createUserToken } from "../lib/helper.ts";
