+++ noj-core/tests/routes/tags.test.ts
@@ -16,7 +16,7 @@ import {
   users,
 } from "../../src/db/schema.ts";
 import { signToken } from "../../src/lib/jwt.ts";
-import { ensureRbacSeeds } from "../../src/services/seed/seed-rbac.ts";
+import { ensureRbacSeeds } from "../../src/domains/system/index.ts";
 
 const hasEnv = !!Deno.env.get("JWT_SECRET");
 const skip = !hasEnv;
