+++ noj-core/tests/services/community.test.ts
@@ -51,8 +51,8 @@ import {
   _resetSystemSettingsForTest,
   initSystemSettings,
   updateSetting,
-} from "../../src/services/system-settings.ts";
-import { ensureRbacSeeds } from "../../src/services/seed/seed-rbac.ts";
+} from "../../src/domains/system/index.ts";
+import { ensureRbacSeeds } from "../../src/domains/system/index.ts";
 
 const actorId = "community-test-user";
 const observerId = "community-test-observer";
