+++ noj-core/tests/services/problem-field-guard.test.ts
@@ -22,16 +22,13 @@ import {
 } from "../../src/db/schema.ts";
 import { AppError, ForbiddenError } from "../../src/lib/errors.ts";
 import { ROOT_USER_ID } from "../../src/lib/constants.ts";
-import { ensureRbacSeeds } from "../../src/services/seed/seed-rbac.ts";
+import { ensureRbacSeeds } from "../../src/domains/system/index.ts";
 import {
   createProblem,
   updateProblem,
 } from "../../src/domains/catalog/index.ts";
 import { importProblemBundle } from "../../src/domains/catalog/index.ts";
-import {
-  resetSetting,
-  updateSetting,
-} from "../../src/services/system-settings.ts";
+import { resetSetting, updateSetting } from "../../src/domains/system/index.ts";
 
 // PGlite 内存数据库始终可用
 const dbAvailable = true;
