+++ noj-core/tests/services/rbac-training.test.ts
@@ -2,7 +2,7 @@ import { assertEquals } from "jsr:@std/assert@^1";
 import { eq } from "drizzle-orm";
 import { getDb, resetDbForTest } from "../../src/db/connection.ts";
 import { permissions, rolePermissions, roles } from "../../src/db/schema.ts";
-import { ensureRbacSeeds } from "../../src/services/seed/seed-rbac.ts";
+import { ensureRbacSeeds } from "../../src/domains/system/index.ts";
 
 await resetDbForTest();
