+++ noj-core/tests/services/system-settings.test.ts
@@ -22,7 +22,7 @@ import {
   maskSecret,
   resetSetting,
   updateSetting,
-} from "../../src/services/system-settings.ts";
+} from "../../src/domains/system/index.ts";
 import {
   _resetEnvSnapshotForTest,
   snapshotEnv,
